Made the field TC a long instead of int

+++ InvoiceApi/UserRegisterDto.cs
@@ -20,6 +20,6 @@ namespace InvoiceApi
         public string Password { get; set; } = String.Empty;
 
         [Required]
-        public int TC { get; set; }
+        public long TC { get; set; }
     }
 }
